Refactor windowed list functions
